Modified the clone method (CLI), and the repo creation (API)

+++ CLI/miniv/Repository.py
@@ -20,24 +20,29 @@ class Create():
         
         if self.__clone_url or self.__create_request:
             '''
-            # Create the repo directory and a .mvcs file in the base dir
+            Create the repo directory and a .mvcs file in the base dir
             '''
+            # get the repo name and the repo path
             repo_name = self.__create_request if self.__create_request else self.__clone_url.split('@')[1].rsplit('/', 2)[-2:][1]
             repo_path = os.path.join(os.getcwd(), repo_name)
+
+            # check if the repository is already cloned or there is directory with the same name
             if not os.path.exists(repo_path):
                 os.mkdir(repo_path, 0o755)
             else:
                 if os.path.exists(repo_path):
                     shutil.rmtree(repo_path)
                 raise Exception("Error, there is already a directory with the same repo name!")
+
+            # create the config folder path
             self.__config_folder = os.path.join(repo_path, ".mvcs")
+
             if not os.path.exists(self.__config_folder):
                 os.mkdir(self.__config_folder, 0o755)
             else:
                 if not os.path.exists(repo_path):
                     shutil.rmtree(repo_path)
                 raise Exception("Error, there already exists a config directory (you are already in a repository)!")
-            self.__UM = UM.UserManagement(self.__config_folder)
         else:
             raise Exception("Error: Wrong input, please clone repository or create a new one!")
 
@@ -67,7 +72,7 @@ class Create():
             user_data['current_branch']= "main"
             user_data['email']         = email
             user_data['id']            = response.json()['user_id']
-            user_data['password']      = self.__UM.encrypt_password(password).decode("utf-8")
+            user_data['password']      = UM.UserManagement.encrypt_password(password).decode("utf-8")
         else:
             raise Exception("Error, wrong credentials, please try again!")
 
@@ -96,27 +101,29 @@ class Create():
         user      = self.__clone_url.split('@')[0]
         host      = self.__clone_url.split('@')[1]
         repo_name = host.rsplit('/', 2)[-2:][1]
+        repo_owner= host.rsplit('/', 2)[-2:][0]
         host      = host.rsplit(':', 1)[0]
-        
-        if self.__UM.check_ssh(host=user, user=user):
-            raise Exception("Error, wrong clone URL!")
 
         # Login user
         user_data = self.login()
-        
+
         ''' 
         # Save the repo data to repo_config.json and 
         # save login data into user_config.json
         # with all the user's data. (in the .mvcs dir) 
         '''
         self.create_user_configuration(user_data)
+        self.__UM = UM.UserManagement(self.__config_folder)
+        
+        if self.__UM.check_ssh(host=user, user=user):
+            raise Exception("Error, wrong clone URL!")
 
         # Creating repo_config.json
         repo_config_file = os.path.join(self.__config_folder, "repo_config.json")
         if not os.path.isfile(repo_config_file):
             try:
                 with open(repo_config_file, 'w+') as f:
-                    response = requests.get(f'http://127.0.0.1:8000/api/repos/data/{repo_name}/')
+                    response = requests.get(f'http://127.0.0.1:8000/api/repos/data/{repo_owner}/{repo_name}/')
                     if response.status_code != 200:
                         raise Exception("Error, requesting repo data failed, please check your clone URL and try again")
                     repo_data = response.json()
@@ -144,16 +151,25 @@ class Create():
         branch_folder = os.path.join(self.__config_folder, 'main')
         commit_folder = os.path.join(branch_folder, os.listdir(branch_folder)[0])
         working_dir = self.__config_folder.split('.mvcs')[0]
-        
-        with tarfile.open(commit_folder) as ccf:
-            ccf.extractall(working_dir)
+        if self.__is_nonempty_tar_file(commit_folder):
+            with tarfile.open(commit_folder) as ccf:
+                ccf.extractall(working_dir)
+                try:
+
+                    path = os.path.join(working_dir, os.listdir(working_dir)[1])
+                    for filename in os.listdir(path):
+                        shutil.move(os.path.join(path, filename), os.path.join(working_dir, filename))
+                    os.rmdir(path)
+                except Exception:
+                    raise Exception("Error happened during downloading the repo!")
+
+    def __is_nonempty_tar_file(self, archive):
+        with tarfile.open(archive, "r") as tar:
             try:
-                path = os.path.join(working_dir, os.listdir(working_dir)[1])
-                for filename in os.listdir(path):
-                    shutil.move(os.path.join(path, filename), os.path.join(working_dir, filename))
-                os.rmdir(path)
-            except Exception:
-                raise Exception("Error happened during downloading the repo!")
+                file_content = tar.getmembers()
+                return len(file_content) > 0
+            except Exception as exc:
+                raise Exception(f"Reading tarfile failed for {archive}")
 
     def create(self):
         '''
